# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ProductRecognitionInput {
  /**
     * Base64 data URL for a JPEG, PNG, GIF, or WebP product photo.
     * @minLength 32
     * @maxLength 6500000
     */
  imageData: string;
  /** @maxLength 4000 */
  ocrText?: string;
}

export interface ProductRecognition {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  brand: string;
}