# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /ai/product-recognition:
    post:
      operationId: recognizeProductPhoto
      tags: [ai]
      summary: Recognize a beauty product from a photo
      description: Uses DeepSeek Vision to suggest a Traditional Chinese product name and brand.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductRecognitionInput"
      responses:
        "200":
          description: Product details recognized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductRecognition"
        "400":
          description: Invalid image input
        "502":
          description: Recognition provider failed
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ProductRecognitionInput:
      type: object
      required:
        - imageData
      properties:
        imageData:
          type: string
          description: Base64 data URL for a JPEG, PNG, GIF, or WebP product photo.
          minLength: 32
          maxLength: 6500000
        ocrText:
          type: string
          maxLength: 4000
    ProductRecognition:
      type: object
      required:
        - name
        - brand
      properties:
        name:
          type: string
          minLength: 1
        brand:
          type: string
          minLength: 1